Implemented a basic way to compute super regions score
Modified the previous structure of the ScoreComputer class by deleting
the two existing functions and creating a function that calculates the
score for a given super region.

Changed the body of "pick_starting_region" so that it doesn't use a
PriorityQueue, but simply the super region with the maximum score

Added a "plan_moves" function in the Bot class to compute a pair of
(source, destination) for deployments and attacks.

Changed the body of "place_armies" to use the value returned by
"plan_moves"

Changed the body of "make_moves" to use the value returned by
"plan_moves"

Added a super region table in the Bot class for ease of access to all
regions inside a super region.

Changed the body of "handle_initial_starting_region" to do nothing
because it might not be needed.

Minor changes in comments.